Update EZGUI to 1.16.5
This imports the EZGUI to 1.16.5
ConfigClient has also been imported, as it was missing.

I've commented out the "onConfigsChanged" function in ConfigClient (lines 189-192), we will need to deal with that when we work on it.

We also need to replace how dimension data is gotten, as we use modid:dimname instead of int dimId in 1.16.5 .